refactor(splash): extract splash logic into custom hook
